Correct sizes on icons and make them square sized
The PNG version of the icon was not square, making it look too wide on
GNOME, on all the Alt-Tab preview, the GNOME-dash and the applications
menu.

Also generated the correct icon sizes for 48x48px and 128x128px. The
previous PNG version was really 120x120px size and saved on GNOME's icon
paths as a 48x48px sized icon.

Corrected the Makefile on the installation part to copy each icon
version to their corresponding location.