Correct symbolic version of the icon
The first part of the N in the symbolic version of the logo was not
being showed as empty or transparent as styled on the official logos.
That wasn't because of the no-fill property for that shape on the SVG
file missing to be set but because GNOME fills any polygon shape when
drawing icons on the top-bar.

The solution was to replace the strikes of that specific shape with
separate lines not forming a polygon but coinciding on their vertices.